requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
frappe-bench==5.22.6
gitdb==4.0.11
GitPython==3.1.43
honcho==1.1.0
idna==3.7
Jinja2==3.1.4
MarkupSafe==2.1.5
python-crontab==2.6.0
python-dateutil==2.9.0.post0
requests==2.32.3
semantic-version==2.8.5
six==1.16.0
smmap==5.0.1
urllib3==2.2.2
